refactor: split tabs from App.tsx into their own components
